+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Script : Add `.script_name` as an alias for `.name`
 - Stage : Add safety checking to *Card-related methods + error for invalid string identifiers
 - TOML : Add `new_config`
+- Util : Add `get_namespace`
 
 ### v0.1.37
 - Internal changes